+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
  - Textures are now inside submodules (for example `Texture2d` is in `texture::texture2d::Texture2d`) and reexported from `texture`.
  - Added `Context::flush()` and `Context::finish()`. Deprecated `Context::synchronize`.
+ - Added `TextureAnyMipmap::raw_upload_from_pixel_buffer`.
 
 ## Version 0.8.2 (2015-07-29)